+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 * Added json parser
+* Fixed automatic variable translation bugs
+* Fixed typescript dumper strings
 
 ## 0.1.1